Jetpack Compose Conversion
Attempted to get a lot of the old View-based code converted to a Jetpack Compose based Activity for easier coding experience. This came with a lot of help from ChatGPT. I made use of it to help me with numerous dependency and gradle issues as well as imports. I also had ChatGPT get a good structure for the layout up in place and connected to our existing ViewModels correctly. I ended up Combining the two view models for easier use.