File Renaming Implemented
Using the help of ChatGPT, I was able to find a clean enough way to allow file naming with our current setup. By simply adding some methods and queries to our RoomDB, DAO, Repository, etc... we are now able to send a new name from the View and have it correctly update both on the disk and on the UI. It also checks for duplicate names so as not to cause additional issues for users. ChatGPT was mainly a big help for the View part of this implementation. I was able to have it setup an Alert Dialogue that would prompt the user for the new name and show error messages when needed.