Code/UI/Documentation Cleanup
Ran through the entire codebase and cleaned it up with the help of ChatGPT. I had it ensure that we were making smart, logical choices and generate concise comments for all the methods we've created. I also tweaked the UI to make sure it had good UX decisions. This project is now nearing its final version and ready for semester end.